+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
  - lti_provider django application
  - install `django-machina` forum in the `sandbox` project
  - custom user model, auth backend and LTI handlers in `ashley`
+ - standalone LTI consumer to test ashley in development
 
 ### Changed